requirements.txt
appdirs==1.4.4
asgiref==3.3.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
distlib==0.3.1
Django==3.1.4
django-tables2==2.3.3
django-crispy-forms==1.10.0
django-extensions==3.1.0
django-background-tasks==1.2.5
django-compat==1.0.15
filelock==3.0.12
gTTS==2.2.1
idna==2.10
pygame==2.0.0
pytz==2020.4
requests==2.25.0
schedule==0.6.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
virtualenv==20.2.2
